backend/requirements.txt
flask
flask-cors
numpy
pillow
segyio
gunicorn